Move header constants to shared-core
